requirements.txt
bcrypt==3.1.7
Cerberus==1.3.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-API==2.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
python-dotenv==0.13.0
requests==2.24.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
